1) добавлены настройки будильников 2) web странца alarm.html редактирования настроек будильников 3) шаблон alarm.ttt содержащий инфо о будильнике 4) переход на uint16_t вместо uint8_t - оптимизация по коду 5) общий буфер fileData для загрузки файлов, гипотетически возможна порча данных при одновременном использовании, на пример обновление погоды и web страницы

+++ LilyGO-TTGO/T-Display/ESP32_I2C_Radio/rda5807m.ino
@@ -42,7 +42,7 @@ uint16_t radioGetChannel() {
   return reg.CHAN + RADIO_BAND_MIN;
 }
 
-void radioSetVolume(uint8_t value) {
+void radioSetVolume(uint16_t value) {
   debug_printf("Volume = %d\r\n", value);
 
   xSemaphoreTake(_mutexRadio, portMAX_DELAY);
